requirements.txt
Flask
numpy
opencv-python-headless
tensorflow
google-cloud-storage
python-dotenv
gunicorn